refactor(frontend): modularize settings screens with organized widget components
